moved the downloading of vcloud jars to vcloud-benchmark.py
