+++ contrib/vcloud-benchmark.py
@@ -10,14 +10,48 @@
 import logging
 import os
 import sys
+import urllib.request
+import subprocess
 
 sys.dont_write_bytecode = True  # prevent creation of .pyc files
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))
 
 from vcloud.vcloudbenchmarkbase import VcloudBenchmarkBase  # noqa E402
+from vcloud import vcloudutil  # noqa E402
 from benchexec import __version__  # noqa E402
 import benchexec.benchexec  # noqa E402
 import benchexec.tools  # noqa E402
+
+_ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "vcloud"))
+IVY_JAR_NAME = "ivy-2.5.0.jar"
+IVY_PATH = os.path.join(_ROOT_DIR, "lib", IVY_JAR_NAME)
+IVY_DOWNLOAD_URL = "https://www.sosy-lab.org/ivy/org.apache.ivy/ivy/" + IVY_JAR_NAME
+
+
+def download_required_jars(config):
+    # download ivy if needed
+    if not os.path.isfile(IVY_PATH):
+        # let the process exit if an exception occurs.
+        urllib.request.urlretrieve(IVY_DOWNLOAD_URL, IVY_PATH)  # noqa S310
+
+    # prepare command
+    cmd = ["java", "-jar", "lib/" + IVY_JAR_NAME]
+    cmd += ["-settings", "lib/ivysettings.xml"]
+    cmd += ["-dependency", "org.sosy_lab", "vcloud", "0.+"]
+    cmd += ["-confs", "runtime", "-mode", "dynamic", "-refresh"]
+    if not config.debug:
+        # In normal mode, -warn is good (no output by default, only if sth. is wrong).
+        # In debug mode, the default Ivy output seems fine (-verbose and -debug would
+        # be too verbose).
+        cmd += ["-warn"]
+    cmd += ["-retrieve", "lib/vcloud-jars/[artifact](-[classifier]).[ext]"]
+
+    # install vcloud jar and dependencies
+    subprocess.run(
+        cmd,
+        cwd=_ROOT_DIR,
+        shell=vcloudutil.is_windows(),  # noqa: S602
+    )
 
 
 class VcloudBenchmark(VcloudBenchmarkBase):
@@ -26,7 +60,13 @@ class VcloudBenchmark(VcloudBenchmarkBase):
     """
 
     def load_executor(self):
+        download_required_jars(self.config)
+
         import vcloud.benchmarkclient_executor as executor
+
+        executor.set_vcloud_jar_path(
+            os.path.join(_ROOT_DIR, "lib", "vcloud-jars", "vcloud.jar")
+        )
 
         logging.debug(
             "This is vcloud-benchmark.py (based on benchexec %s) "
