+++ contrib/vcloud/benchmarkclient_executor.py
@@ -11,8 +11,6 @@ import logging
 import os
 import shutil
 import subprocess
-import urllib.request
-
 import benchexec.tooladapter
 import benchexec.util
 from . import vcloudutil
@@ -27,13 +25,12 @@ DEFAULT_CLOUD_CPUMODEL_REQUIREMENT = ""  # empty string matches every model
 
 STOPPED_BY_INTERRUPT = False
 
-_ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__)))
-
 _JustReprocessResults = False
 
-IVY_JAR_NAME = "ivy-2.5.0.jar"
-IVY_PATH = os.path.join(_ROOT_DIR, "lib", IVY_JAR_NAME)
-IVY_DOWNLOAD_URL = "https://www.sosy-lab.org/ivy/org.apache.ivy/ivy/" + IVY_JAR_NAME
+
+def set_vcloud_jar_path(p):
+    global vcloud_jar
+    vcloud_jar = p
 
 
 def init(config, benchmark):
@@ -52,32 +49,6 @@ def init(config, benchmark):
 
 def get_system_info():
     return None
-
-
-def download_required_jars(config):
-    # download ivy if needed
-    if not os.path.isfile(IVY_PATH):
-        # let the process exit if an exception occurs.
-        urllib.request.urlretrieve(IVY_DOWNLOAD_URL, IVY_PATH)  # noqa S310
-
-    # prepare command
-    cmd = ["java", "-jar", "lib/" + IVY_JAR_NAME]
-    cmd += ["-settings", "lib/ivysettings.xml"]
-    cmd += ["-dependency", "org.sosy_lab", "vcloud", "0.+"]
-    cmd += ["-confs", "runtime", "-mode", "dynamic", "-refresh"]
-    if not config.debug:
-        # In normal mode, -warn is good (no output by default, only if sth. is wrong).
-        # In debug mode, the default Ivy output seems fine (-verbose and -debug would
-        # be too verbose).
-        cmd += ["-warn"]
-    cmd += ["-retrieve", "lib/vcloud-jars/[artifact](-[classifier]).[ext]"]
-
-    # install vcloud jar and dependencies
-    subprocess.run(
-        cmd,
-        cwd=_ROOT_DIR,
-        shell=vcloudutil.is_windows(),  # noqa: S602
-    )
 
 
 def execute_benchmark(benchmark, output_handler):
@@ -101,8 +72,6 @@ def execute_benchmark(benchmark, output_handler):
             }
         )
 
-        download_required_jars(benchmark.config)
-
         # start cloud and wait for exit
         logging.debug("Starting cloud.")
         if benchmark.config.debug:
@@ -111,12 +80,12 @@ def execute_benchmark(benchmark, output_handler):
             logLevel = "INFO"
         # heuristic for heap size: 100 MB and 100 kB per run
         heapSize = benchmark.config.cloudClientHeap + numberOfRuns // 10
-        lib = os.path.join(_ROOT_DIR, "lib", "vcloud-jars", "vcloud.jar")
+        # vcloud_jar is to be set by the calling script, i.e., (vcloud-)benchmark.py
         cmdLine = [
             "java",
             "-Xmx" + str(heapSize) + "m",
             "-jar",
-            lib,
+            vcloud_jar,
             "benchmark",
             "--loglevel",
             logLevel,
